requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
xgboost>=1.7.0
joblib>=1.2.0
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.0
flask>=2.3.0
flask-cors>=4.0.0
opencv-python>=4.8.0
mediapipe>=0.10.0
pillow>=10.0.0
werkzeug>=2.3.0

